requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.8
python-docx==1.1.0
pypdf==4.0.2
gradio==4.19.2 